requirements.txt
pandas
pip
ipykernel
ipywidgets
jupyter
tqdm

# pip install web3 --only-binary :all:
web3>=5.29.0
py-solc-x >= 1.1.1
eth-tester >= 0.6.0b6
py-evm


